integratie met folie

+++ UV-sensor.h
@@ -1,9 +1,9 @@
-#include <TinyWireM.h>
+#include <Wire.h>
 #define SI1145_ADDR 0x60
 
 typedef struct {
-    uint16_t vis;
-    uint16_t uv;
+    unsigned int vis;
+    unsigned int uv;
 } SI1145_value;
 
 int SI1145_init_sensor();
